# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Warehouse Designer API",
    "description": "Authoritative layout compiler and persistence for the 3D warehouse designer. The TypeScript compiler in packages/layout-core is the interactive mirror of the compiler used here; both are pinned by fixtures/layout-conformance.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/warehouses/{warehouse_id}/skus": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "List Skus",
        "operationId": "list_skus_api_warehouses__warehouse_id__skus_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkuOut"
                  },
                  "title": "Response List Skus Api Warehouses  Warehouse Id  Skus Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Create Sku",
        "operationId": "create_sku_api_warehouses__warehouse_id__skus_post",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkuIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkuOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skus/{sku_id}": {
      "put": {
        "tags": [
          "inventory"
        ],
        "summary": "Update Sku",
        "description": "Edit a SKU in place.\n\nThis does **not** re-validate existing placements. Shrinking a SKU cannot invalidate\nanything, but growing one can, and silently deleting inventory because a dimension\nchanged would be worse than leaving a placement that a later check flags. The\nplacement list reports each bin's utilisation, so an over-capacity bin is visible.",
        "operationId": "update_sku_api_skus__sku_id__put",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sku Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkuIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkuOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "inventory"
        ],
        "summary": "Delete Sku",
        "operationId": "delete_sku_api_skus__sku_id__delete",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sku Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/placements": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "List Warehouse Placements",
        "operationId": "list_warehouse_placements_api_warehouses__warehouse_id__placements_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlacementOut"
                  },
                  "title": "Response List Warehouse Placements Api Warehouses  Warehouse Id  Placements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/utilization": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Utilization",
        "operationId": "get_utilization_api_warehouses__warehouse_id__utilization_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilizationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bins/{bin_id}/placements": {
      "get": {
        "tags": [
          "inventory"
        ],
        "summary": "Get Bin Placements",
        "operationId": "get_bin_placements_api_bins__bin_id__placements_get",
        "parameters": [
          {
            "name": "bin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Bin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinPlacementsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Create Placement",
        "description": "Place a SKU in a bin, revalidating on the server.\n\nThe client checks the same thing before the drop, so a 200 here means both sides\nagreed and a 422 means the codes should match what the client computed.",
        "operationId": "create_placement_api_bins__bin_id__placements_post",
        "parameters": [
          {
            "name": "bin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Bin Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlacementIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/placements/{placement_id}": {
      "delete": {
        "tags": [
          "inventory"
        ],
        "summary": "Delete Placement",
        "operationId": "delete_placement_api_placements__placement_id__delete",
        "parameters": [
          {
            "name": "placement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Placement Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/warehouses/{warehouse_id}/placements/bulk": {
      "post": {
        "tags": [
          "inventory"
        ],
        "summary": "Bulk Create Placements",
        "description": "Fill many bins at once, with a verdict per bin rather than all-or-nothing.",
        "operationId": "bulk_create_placements_api_warehouses__warehouse_id__placements_bulk_post",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Warehouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPlacementIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPlacementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BinPlacementsOut": {
        "properties": {
          "binId": {
            "type": "string",
            "format": "uuid",
            "title": "Binid"
          },
          "binCode": {
            "type": "string",
            "title": "Bincode"
          },
          "capacityM3": {
            "type": "number",
            "title": "Capacitym3"
          },
          "maxWeightKg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxweightkg"
          },
          "usedVolumeM3": {
            "type": "number",
            "title": "Usedvolumem3"
          },
          "usedWeightKg": {
            "type": "number",
            "title": "Usedweightkg"
          },
          "remainingVolumeM3": {
            "type": "number",
            "title": "Remainingvolumem3"
          },
          "remainingWeightKg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remainingweightkg"
          },
          "utilization": {
            "type": "number",
            "title": "Utilization"
          },
          "placements": {
            "items": {
              "$ref": "#/components/schemas/PlacementOut"
            },
            "type": "array",
            "title": "Placements"
          }
        },
        "type": "object",
        "required": [
          "binId",
          "binCode",
          "capacityM3",
          "maxWeightKg",
          "usedVolumeM3",
          "usedWeightKg",
          "remainingVolumeM3",
          "remainingWeightKg",
          "utilization",
          "placements"
        ],
        "title": "BinPlacementsOut",
        "description": "A bin and everything in it, with the capacity arithmetic already done.\n\n`remaining*` is computed from the *other* placements, so the client can render a\ndrag ghost from the same numbers the server checks rather than re-deriving them."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlacementIn": {
        "properties": {
          "skuId": {
            "type": "string",
            "format": "uuid",
            "title": "Skuid"
          },
          "qty": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Qty"
          }
        },
        "type": "object",
        "required": [
          "skuId",
          "qty"
        ],
        "title": "PlacementIn"
      },
      "PlacementOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "binId": {
            "type": "string",
            "format": "uuid",
            "title": "Binid"
          },
          "binCode": {
            "type": "string",
            "title": "Bincode"
          },
          "skuId": {
            "type": "string",
            "format": "uuid",
            "title": "Skuid"
          },
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "qty": {
            "type": "integer",
            "title": "Qty"
          },
          "volumeUsedM3": {
            "type": "number",
            "title": "Volumeusedm3"
          },
          "weightUsedKg": {
            "type": "number",
            "title": "Weightusedkg"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "binId",
          "binCode",
          "skuId",
          "sku",
          "qty",
          "volumeUsedM3",
          "weightUsedKg"
        ],
        "title": "PlacementOut"
      },
...
      "SkuIn": {
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Sku"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "widthM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Widthm"
          },
          "heightM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Heightm"
          },
          "depthM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Depthm"
          },
          "weightKg": {
            "type": "number",
            "minimum": 0.0,
            "title": "Weightkg"
          },
          "stackable": {
            "type": "boolean",
            "title": "Stackable",
            "default": true
          },
          "rotatable": {
            "type": "boolean",
            "title": "Rotatable",
            "default": true
          },
          "hazmat": {
            "type": "boolean",
            "title": "Hazmat",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "sku",
          "widthM",
          "heightM",
          "depthM",
          "weightKg"
        ],
        "title": "SkuIn"
      },
      "SkuOut": {
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Sku"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "widthM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Widthm"
          },
          "heightM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Heightm"
          },
          "depthM": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Depthm"
          },
          "weightKg": {
            "type": "number",
            "minimum": 0.0,
            "title": "Weightkg"
          },
          "stackable": {
            "type": "boolean",
            "title": "Stackable",
            "default": true
          },
          "rotatable": {
            "type": "boolean",
            "title": "Rotatable",
            "default": true
          },
          "hazmat": {
            "type": "boolean",
            "title": "Hazmat",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "placedQty": {
            "type": "integer",
            "title": "Placedqty",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "sku",
          "widthM",
          "heightM",
          "depthM",
          "weightKg",
          "id"
        ],
        "title": "SkuOut"
      },
      "UtilizationOut": {
        "properties": {
          "warehouseId": {
            "type": "string",
            "format": "uuid",
            "title": "Warehouseid"
          },
          "binCount": {
            "type": "integer",
            "title": "Bincount"
          },
          "utilizedBins": {
            "type": "integer",
            "title": "Utilizedbins"
          },
          "emptyBins": {
            "type": "integer",
            "title": "Emptybins"
          },
          "entries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "warehouseId",
          "binCount",
          "utilizedBins",
          "emptyBins",
          "entries"
        ],
        "title": "UtilizationOut",
        "description": "Heatmap input: utilisation per bin, and the bins that hold nothing."
      },
...
    }
  }
}